base: add DefaultTransport; update all New() functions
Change all New() provider functions to accept a wrapper. While unusual
API design, it's the design that provides the cleanest interface to setup HTTP
recordings and playback.

It is still possible to late-hook like it was done before but this won't
work with runtime model selection that will be added in a follow up.