Update llamacppsrv and set version for ollamasrv
It's a bit awkward because the current good version is v0.9.1-rc0 but it works...

Compared to v0.7.0, v0.9.1-rc0 is really slow to start and shutdown. A
bummer.

Tried to move the temporary files under testdata/tmp/ but it's still not
caching tests, there's something wrong with the test caching that I need
to figure out.