adapters: ProviderGenThinking remove SkipJSON, require start tag
This unblocks decoding of JSON responses. This makes it more versatile.
On the other hand, the start tag is now required. It could be made
optional if required in a follow up but then it'd be hard to also
support JSON responses.

Fix error message when there's content before a thinking tag.