scoreboard: add many test data files
Recreate them instead of taking the old ones from internaltest. Include
the bash shell scripts to reconstruct them all.

Will switch scoreboard to these files in the next commit, since the file
have to exist in the github hosted repository first so the URL version
of the API calls work.