gemini: implement GenAsync and confirm Veo 2 works
I realize a lot of features for Veo 3 are only exposed in the Vertex AI
API and not on the Gemini API. I may have to switch sooner than
expected.

I also realized that the Vertex AI API supports API keys, so it may be
easier than expected.

Because scoreboard doesn't implement video output testing yet, there's
no video output tag on the Gemini provider. The annoyance is that each
test run cost 5$ so it's not fun to re-run the test. I'll revisit when I
can keep costs under control.