gemini: stop leaking API key (again)
When fetching a Veo 2 generated file, there's an HTTP 302 redirect
happening. The redirect includes the API key in the response HTTP
headers (!!)

Use the occasion to remove the code for the "key" query argument in the
URL, I switched to using "X-Goog-Api-Key". This removes a good deal of
dead code.