genai: enforce alternating messages
Some models, like gemma 3, requires roles to be alternating. Now that
multi-content messages are supported on all providers, it's better
to only have one way of doing things. I prefer to do this change upfront
because it's quite a breaking change and would be annoying to do later
if there's more users.