openaichat: fix thinking processing
OpenAI's chat api doesn't ever return thinking tokens but does return
the number of reasoning tokens generated. The client implementation used
to send whitespaces to mark that but this is really a bad idea. Now that
Usage.ReasoningTokens is returned to the user, leverage this to inject
fake thinking tokens but only to the scoreboard, so that the scoreboard
remains valid but the data returned to real users is clean.