scoreboard: Huge refactoring
- smoke: Extract scoreboard smoke test code into a new package smoke.
- smoke: Rename CreateScenario to Run.
- smoketest: Rename AssertScoreboard to Run.
- smoketest: rename GetClient to ProviderFactory.
- genai: Merge scoreboard.ProviderScoreboard into genai.Provider. This
  enforces all providers have a scoreboard. This is the way.
- scoreboard: keep testdata/ and export TestdataFiles to be the
  canonical files that describes the supported modalities. This may
  belong to package smoke but this requires more work to redo the
  recording because many of these have the path embedded in.

This greatly reduce the need for Unwrap(), which is important.

I decided against moving the symbols in scoreboard into genai to reduce
noise. The drawback is that I have to alias Modality to not cause a
circular dependency. I currently believe it's the right choice but I'm
open to be convinced otherwise.

Fairly breaking change.