cmd/scoreboard: use a different emoji for stream
When it has partial support, use 🤏.

This is not great, I'll have to redo the columns. Two options I'm
thinking about:
- Use two values per column, e.g. "✅/❌".
- Use two rows, one for sync, one for stream.

The second has the advantage of being more explicit when stream is not
supported, e.g. bfl.

In both case, I'll probably want to move finish reason and token usage
to a new column.