smoke/smoketest: force the scoreboard to be in order
This forced me to revisit some of the selected models, which is good. I
plan this to be a regular task as the field moves fast!

The linter is still broken because it depends on having an access token.
The only fix is to add the selected models to the scoreboard. It's a bit
annoying but I think it's the right call. This will be done in a follow
up.