smoke: stop considering tool non-required as flaky
Initially I considered any provider that doesn't support
ToolCallRequired as flaky. Now that it is a different Functionality
value, it is okay to not consider the model as flaky if it succeeds on
first try without ToolCallRequired.

The fact that ToolCallRequired is not supported is a big problem but
that's separate from the fact that tool calling works well or not.