base&ollama: GenStreamRaw: use iterators
GenStream still use channels internally because of
ProcessStreamPackets(). This will be changed in a follow up.

Make GenStreamRaw have an internal buffer of 32 packets. This should
make sure that the HTTP stream is processed enough in advance so that
the underlying TCP connection doesn't stall even if the packet
processing has some hickups, like printing to stdout.